pyproject.toml
[project]
name = "inlive-python-sdk"
version = "0.1.0"
description = "Python SDK for inLive platform, including RTC and recording modules."
authors = [
    { name = "Your Name", email = "your@email.com" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiortc>=1.11.0",
    "aioice>=0.7.5",
    "httpx>=0.24.0",
    "pyee>=9.0.0",
    "pytest>=8.3.5",
    "requests>=2.32.3",
    "fastapi>=0.115.12",
    "python-dotenv>=1.0.1",
    "pytest-asyncio>=0.24.0",
    "numpy>1.24.4",
]

[project.urls]
Homepage = "https://github.com/inlivedev/inlive-python-sdk"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
